lots of work on clockwork sms api

+++ src/wbs/clients/apn/chat/broadcast/console/ChatBroadcastSendAction.java
@@ -659,7 +659,7 @@ class ChatBroadcastSendAction
 						form.prefix (),
 						form.message ());
 
-				int messageLength =
+				long messageLength =
 					Gsm.length (
 						messageString);
